refactor: integrate issue #34 and #29 tests into existing test suite
- Remove separate test files (repro-issue-34.test.ts, issue-34-integration.test.ts)
- Remove example file (site/examples/issue-34-reproduction.page.tsx)
- Add Issue #34 test case to existing mergeCollinearTraces.test.ts:
  * Tests three fragmented collinear segments merge into one (0 to 10)
- Add Issue #29 test case to existing mergeCollinearTraces.test.ts:
  * Tests overlapping segments on same net merge correctly (0 to 10 with overlap)
- Keep only production code changes in TraceCleanupSolver.ts
- Maintain project's native testing patterns
